+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # LPC Language Services Changelog
 
+## 1.1.35
+
+-   Added support for `directory_exists` efun use in master apply `get_include_path`.
+
 ## 1.1.34
 
 -   [Type predicates are not working on object declarations #238](https://github.com/jlchmura/lpc-language-server/issues/238)